requirements.txt
streamlit
requests
openai
tensorflow
folium
sklearn
numpy
pandas
pydeck
jinja